### 2021-05-06 #3253 @koaning
No rush.

One question though, this is a package that I maintain personally which means that I don't have a `affiliations` entry. Should I just put in my name here?
