### 2021-05-20 #3253 @kyleniemeyer
@koaning I think you still may need the affiliations field, even if you leave it empty

cc @openjournals/dev just checking, is that correct?
