### 2021-05-20 #3253 @arfon
> you can't leave it empty, so you need to put in one that says "independent researcher" or "self" etc...

Correct.
