### 2021-05-20 #3253 @koaning
Still strange that the compiler did not complain about it.
