### 2021-06-21 #3253 @jbytecode
@koaning I notice that you're the only author on the paper but it looks like there have been some other contributors to the codebase. Could you please clarify?
