### 2021-06-21 #3253 @koaning
There are three contributors outside of myself.

One contributor fixed a typo:
https://github.com/koaning/human-learn/commit/deace163a8e2f381392c08119ecad6944f48c1d8

Another one contributed a one-liner so that the license would be included in the source distribution:
https://github.com/koaning/human-learn/commit/188d16775393edcafcf77d91b4f2491f79a3e3d5

Finally, there's one contributor who added a dataset to the repository. That's a welcome PR, but it's very much a utility function that doesn't touch the internals of the package.
https://github.com/koaning/human-learn/commits?author=kayhoogland

I don't mind mentioning these contributors on the paper, but since none of the other contributions touched the internal components (what the package is really about) it didn't occur to me to mention them in the paper directly.
