### 2021-06-21 #3253 @jbytecode
@jankatins hi! I am pinging you just because you are registered as an expert on Python and visualizing on the JOSS reviewer list. Would you willing to assist in reviewing this submission?
