### 2021-06-21 #3253 @jbytecode
> I'd be happy to review this but I can't do it until mid-July unfortunately.

@desilinguist thank you! The reviewing process will start whenever we assign a second viewer. Meanwhile, you can have a look at the reviewing guide in https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
