### 2021-06-21 #3253 @jbytecode
@koaning while we are seeking a proper reviewer team, please consider following items:

- The paper does not compare the software with the previous one(s). In JOSS, this kind of discussion is generally handled in a `State of the field` section.
- It is not usual for an article not to have a reference list. I think, when you add comparisons in State of the field section, you will use some citations.
