### 2021-06-29 #3253 @jbytecode
Hi @kellieotto, would you be willing to assist in reviewing this submission for JOSS?

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "Human-Learn". You can find more information at the top of this Github issue.
