### 2021-07-02 #3448 @whedon
[REVIEW]: Human-Learn
**Submitting author:** @koaning (<a href="http://orcid.org/0000-0003-0845-4528">Vincent Warmerdam</a>)
**Repository:** https://github.com/koaning/human-learn/
**Branch with paper.md** (empty if default branch):
**Version:** 0.3.0
