### 2021-07-02 #3448 @jbytecode
Dear @desilinguist  and @ahurriyetoglu ,

This is the review issue for the submission. There are 20 checkbox items for each reviewer. Please check the boxes during your review and make comments here. Since the reviewing is progressive, you can comment minor changes during the process so the authors can add/delete/correct changes simultaneously. You can also open new issues in the submitted repository and put links here. Here is the reviewing guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Please do not hesitate to ask for help in any stage.

Thank you in advance for your contributions to JOSS!
