### 2021-07-11 #3448 @desilinguist
This package looks really cool and something that I would likely check out myself for some work that I am doing! Excellent work! The documentation is top-notch. I only have some minor suggestions:

1. Please clearly list the dependencies in the GitHub README, if possible. This makes it much easier to figure out what's needed than looking at requirements.txt.

2. I didn't see API documentation for the various classes in the official docs.

3. It's clear from the code that there are tests but it would if the README described how to run them. And this leads me to ...

4. There is no section in the README or in the docs about inviting contributions from the community or how someone who might be interested in this would contribute.

And then one final comment that's more selfish, it would be really nice if there was a conda package in addition to one on PyPI but that's totally optional, of course :)

In general, great work and love the clear examples and walkthroughs!
