### 2021-07-11 #3448 @koaning
Happy to hear you like it!

1. Done.
2. Could you give an example of the classes you found were missing?
3. Done.
4. Done.

I might consider conda if there's a clear documented method for me to deploy there. Got any links with tutorials?
