### 2021-07-11 #3448 @desilinguist
@koaning I was looking for API documentation that looked something like [this](https://skll.readthedocs.io/en/latest/api.html). If you are using Sphinx, this can be generated automatically using the [autodoc directive](https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html). When I am using a library, I find such an API reference quite useful.

As for deploying to conda, here are a couple of pointers that will prove useful:

1. Using [conda-build](https://docs.conda.io/projects/conda-build/en/latest/#) to create a conda package.

2. Once you have built the package, you can sign up for a free anaconda.org account to distribute it via a custom channel or you can [contribute the package](https://conda-forge.org/#add_recipe) to the community conda-forge channel.

FYI, I have a lot of experience with building conda packages so let me know if you need any additional help.
