### 2021-07-11 #3448 @koaning
I'm using the `mkdocs` plugin together with `mkdocstrings` which should auto-generate the API docs for every class/function that has a docstring. You can find it [here](https://koaning.github.io/human-learn/api/classification.html).
