### 2021-07-12 #3448 @jbytecode
@koaning could you please clarify why the submitted paper has no citations? It is not usual not to have citations as the author(s) are not the whole inventors of the subject at hand. Please add citations and reference entities possibly on 1) the methods that are implemented 2) the technology which is built on 3) the terminology used before.
