### 2021-07-12 #3448 @koaning
That's odd. I added a [paper.bib](https://github.com/koaning/human-learn/blob/master/paper.bib) to the main repo but it's not being picked up. Lemme investigate that.
