### 2021-07-12 #3448 @jbytecode
> That's odd. I added a [paper.bib](https://github.com/koaning/human-learn/blob/master/paper.bib) to the main repo but it's not being picked up, despite being listed in the frontmatter [here](https://raw.githubusercontent.com/koaning/human-learn/master/paper.md). Any clue why it's not being picked up?

it is just because entries in the bibtex file are not cited in the main paper using `[@entrylabel]`
