+++ paper.bib
@@ -37,3 +37,26 @@
   year      = {2013},
   pages = {108--122},
 }
+
+@incollection{NIPS2017_7062,
+  title = {A Unified Approach to Interpreting Model Predictions},
+  author = {Lundberg, Scott M and Lee, Su-In},
+  booktitle = {Advances in Neural Information Processing Systems 30},
+  editor = {I. Guyon and U. V. Luxburg and S. Bengio and H. Wallach and R. Fergus and S. Vishwanathan and R. Garnett},
+  pages = {4765--4774},
+  year = {2017},
+  publisher = {Curran Associates, Inc.},
+  url = {http://papers.nips.cc/paper/7062-a-unified-approach-to-interpreting-model-predictions.pdf}
+}
+
+@inproceedings{lime,
+  author    = {Marco Tulio Ribeiro and
+               Sameer Singh and
+               Carlos Guestrin},
+  title     = {"Why Should {I} Trust You?": Explaining the Predictions of Any Classifier},
+  booktitle = {Proceedings of the 22nd {ACM} {SIGKDD} International Conference on
+               Knowledge Discovery and Data Mining, San Francisco, CA, USA, August
+               13-17, 2016},
+  pages     = {1135--1144},
+  year      = {2016},
+}
