### 2021-07-12 #3448 @koaning
This should now be fixed. I also added a sentence about LIME and SHAP. They had a role to play in developing this tool so it's fair to mention these tools as well.
