### 2021-07-12 #3448 @jbytecode
@koaning please update missing DOI's as whedon suggests.
