### 2021-07-12 #3448 @jbytecode
@koaning the correct use of citations is using the brackets like `[@citationlabel]`, not parenthesis like `(@citationlabel)`. Please update the correct form later.
