+++ paper.bib
@@ -53,6 +53,7 @@
   author    = {Marco Tulio Ribeiro and
                Sameer Singh and
                Carlos Guestrin},
+  doi       = {10.18653/v1/n16-3020},
   title     = {"Why Should {I} Trust You?": Explaining the Predictions of Any Classifier},
   booktitle = {Proceedings of the 22nd {ACM} {SIGKDD} International Conference on
                Knowledge Discovery and Data Mining, San Francisco, CA, USA, August
