### 2021-07-18 #3448 @ahurriyetoglu
@koaning Thanks for this cool software. I am interested and have worked on combining rules and machine learning. I started the review by reading the paper from the PDF I found in this thread, generated 6 days ago. My points:

1- The English should be improved. Some examples:
"There’s" -> "There is" There are many cases like this.
python -> Python
"also host a" -> also hosts
"(Kluyver et al. (2016))" -> (Kluyver et al., 2016)
2- I do not see the list of the dependencies on the repo. Where is the requirements.txt file?
3- The paper is not is not self contained. A reader would not get an overview of the tool by reading this paper. The functionality should be summarized in the paper.
4- There is not any examples in the paper. Providing links to external tutorials etc. is not enough.
5- State-of-the-art should be explained. A mention of Snorkel could be helpful.

Please let me know if you think I can provide more details about the points above. I am looking forward to reading the next version.

Best wishes,

Ali
