### 2021-07-19 d8bc7b1
Update paper.md
