+++ paper.md
@@ -33,7 +33,7 @@ We started wondering if we might have lost something in this transition. Machine
 
 At the same time, it's also true that many classification problems can be done by natural intelligence. This package aims to make it easier to turn the act of exploratory data analysis into a well-understood model. These "human" models are very explainable from the start. If nothing else, they can serve as a simple benchmark representing domain knowledge which is a great starting point for any predictive project.
 
-The library features components to easily turn python functions into scikit-learn compatible components [@sklearn_api]. Both the `FunctionClassifier` and the `FunctionRegressor` can turn python functions into grid-searchable components. Human-learn also hosts a suite of interactive widgets, made with Bokeh, that might help construct models from Jupyter as well.
+The library features components to easily turn python functions into scikit-learn compatible components [@sklearn_api]. Both the `FunctionClassifier` and the `FunctionRegressor` can turn python functions into grid-searchable components. Human-learn also hosts interactive widgets, made with Bokeh, that might help construct models from Jupyter as well.
 
 ![](docs/screenshot.png)
 
