+++ paper.bib
@@ -61,3 +61,18 @@
   pages     = {1135--1144},
   year      = {2016},
 }
+
+@article{snorkel,
+   title={Snorkel},
+   volume={11},
+   ISSN={2150-8097},
+   url={http://dx.doi.org/10.14778/3157794.3157797},
+   DOI={10.14778/3157794.3157797},
+   number={3},
+   journal={Proceedings of the VLDB Endowment},
+   publisher={VLDB Endowment},
+   author={Ratner, Alexander and Bach, Stephen H. and Ehrenberg, Henry and Fries, Jason and Wu, Sen and Ré, Christopher},
+   year={2017},
+   month={Nov},
+   pages={269–282}
+}
