### 2021-07-19 #3448 @koaning
1. Fair! I just ran Grammarly over it. The `(Kluyver et al., 2016)` example is generated by bibtex though.
2. The dependencies are listed in the `setup.py` file.
3-4. Added the code for the function classifiers, this should give a proper overview of what the library is about. Also added the example for the fraud case.
5. I guess I could mention that project. The goal of this project isn't to be state of the art at all, rather, to just remain incredibly simple to use.
