### 2021-07-19 #3448 @jbytecode
@koaning Whatever the message contains, the whedon commands should be on the first line.
