### 2021-07-22 #3448 @jbytecode
@koaning the example code added in the paper overfits the page so
- please refactor the code so that it doesn't overflow the width of the paper.
- please use a space before and after a binary operator, so the code will be more readable, for example `a=b` become `a = b`.
- please test your pdf using https://whedon.theoj.org/ before trig `whedon generate pdf` here.
