+++ paper.md
@@ -54,7 +54,7 @@ Besides the `FunctionClassifier`, the library also features a `FunctionRegressor
 
 ```python
 from sklearn.model_selection import GridSearchCV
-from sklearn.metrics import precision_score, recall_score, accuracy_score, make_scorer
+from sklearn.metrics import *
 
 # The GridSearch object can now "grid-search" over this model.
 grid = GridSearchCV(mod,
