### 2021-07-23 #3448 @koaning
The code no longer overflows. I'm not 100% what your comment on binary operators is referring to though. I've formatted the code using `black` which means that all declarations when defining a function are written via `param=default`. Is this what you're referring to? I tried looking for it but the binary operator `~` is not used anywhere in the paper.
