### 2021-07-27 #3448 @ahurriyetoglu
Thanks @koaning! A couple of other minor points:
1- You can cite the scikit-learn paper and provide a URL to their homepage. (Pedregosa, F., Varoquaux, G., Gramfort, A., Michel, V., Thirion, B., Grisel, O., ... & Duchesnay, E. (2011). Scikit-learn: Machine learning in Python. the Journal of machine Learning research, 12, 2825-2830.)

2- python -> Python, it's -> it is, user the define rules -> user to define rules, snokel -> Snorkel
3- The Figures, code snippets, etc. should have some name and a reference in the text. It is not easy to understand what are the code snippets, what they do, why they are there, etc.

Best wishes,

Ali
