+++ paper.md
@@ -101,7 +101,7 @@ model = InteractiveClassifier(json_desc=clf.data())
 
 This project was developed in my spare time while being employed at Rasa. They have been very supportive of me working on my own projects on the side, and I would like to recognize them for being a great employer.
 
-I also want to acknowledge that I'm building on the shoulders of giants. The popular drawing widget in this library would not have been possible without the wider bokeh [@bokeh], jupyter [@jupyter] and scikit-learn communities.
+I also want to acknowledge that I'm building on the shoulders of giants. The popular drawing widget in this library would not have been possible without the wider bokeh [@bokeh], jupyter [@jupyter] and scikit-learn [@sklearn] communities.
 
 There have also been small contributions on Github from Joshua Adelman, Kay Hoogland, and Gabriel Luiz Freitas Almeida.
 
