### 2021-07-27 #3448 @koaning
1. I already cited the scikit-learn pipeline design in the paper, but I've added a citation in the acknowledgment section using a bibtex citation to the paper.
2. Done.
3. Done.
