### 2021-07-27 #3448 @koaning
Looking at the list on top I figured I should also add installation instructions in the paper. I added it under a new "features" section that picks up right after the statement of need.
