### 2021-07-31 #3448 @ahurriyetoglu
Thanks @koanin. Thanks for the update, it looks better now. Some more points:

1- There's been -> there has been
2- don't -> do not
3- Snokel -> Snorkel
4-"in figure Figure 4" -> in Figure 4
5- Not all figures are mentioned in the text.
6- Code snippets require a numbering and in-text-mentioning as well.
7- The following snippet could be written better in terms of English Language: "Machine learning is a general tool, but it is capable of making bad decisions. Decisions that are very hard to debug too."
8- Please make the "import *" part explicit. In this confusing in this form.

You report the capabilities. Do you have any results and an error analysis of them to report? This could be a use-case. I believe showing performance of this tool will increase its impact a lot!

Best. wishes,
