+++ paper.md
@@ -21,15 +21,15 @@ This package contains scikit-learn compatible tools that make it easier to const
 
 # Statement of need
 
-There's been a transition from rule-based systems to ones that use machine learning. Initially, systems converted data to labels by applying rules.
+There has been a transition from rule-based systems to ones that use machine learning. Initially, systems converted data to labels by applying rules, like in \autoref{fig:rulebased}.
 
 ![Rule Based Systems.\label{fig:rulebased}](docs/examples/rules.png)
 
-Recently, it has become much more fashionable to take data with labels and to use machine-learning algorithms to figure out appropriate rules.
+Recently, it has become much more fashionable to take data with labels and to use machine-learning algorithms to figure out appropriate rules, like in \autoref{fig:mlbased}. 
 
 ![Machine Learning Based Systems.\label{fig:mlbased}](docs/examples/ml.png)
 
-We started wondering if we might have lost something in this transition. Machine learning is a general tool, but it is capable of making bad decisions. Decisions that are very hard to debug too.  Tools like SHAP [@NIPS2017_7062] and LIME [@lime] try to explain why algorithms make certain decisions in hindsight, but even with the benefit of hindsight, it is tough to understand what is happening. 
+We started wondering if we might have lost something in this transition. Machine learning is a general technique, but it's proven to be very hard te debug. This is especially painful when wrong predictions are made.  Tools like SHAP [@NIPS2017_7062] and LIME [@lime] try to explain why algorithms make certain decisions in hindsight, but even with the benefit of hindsight, it is tough to understand what is happening. 
 
 At the same time, it is also true that many classification problems can be done by natural intelligence. This package aims to make it easier to turn the act of exploratory data analysis into a well-understood model. These "human" models are very explainable from the start. If nothing else, they can serve as a simple benchmark representing domain knowledge which is a great starting point for any predictive project.
 
@@ -77,11 +77,11 @@ grid.fit(X, y)
 
 These function-based models can be very powerful because they allow the user the define rules for situations for which there is no data available. In the case of financial fraud, if a child has above median income, this should trigger risk. Machine learning models cannot learn if there is no data but rules can be defined even if, in this case, a child with above median income doesn't appear in the training data. An ideal use-case for this library is to combine rule based systems with machine learning based systems. An example of this is shown in \autoref{fig:tree}. 
 
-![A rule based systems that resorts to ML when rules don't cover the example.\label{fig:tree}](https://koaning.github.io/human-learn/examples/tree.png)
+![A rule based systems that resorts to ML when rules do not cover the example.\label{fig:tree}](https://koaning.github.io/human-learn/examples/tree.png)
 
-This example also demonstrates the main difference between this library and Snokel [@snorkel]. This library offers methods to turn domain knowledge immediately into models, as opposed to labelling-functions.
+This example also demonstrates the main difference between this library and Snorkel [@snorkel]. This library offers methods to turn domain knowledge immediately into models, as opposed to labelling-functions.
 
-Human-learn also hosts interactive widgets, made with Bokeh, that might help construct models from Jupyter as well. An example of a drawn widget is shown below in figure \autoref{fig:draw}.
+Human-learn also hosts interactive widgets, made with Bokeh, that might help construct models from Jupyter as well. An example of a drawn widget is shown below in \autoref{fig:draw}.
 
 ```python
 from hulearn.experimental.interactive import InteractiveCharts
