### 2021-07-31 #3448 @koaning
1. Done
2. Done
3. Done
4. Done
5. Done
6. I tried finding a guide on how to format the code snippets but couldn't find anything in [this guide](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography). Is there a guide elsewhere?
7. Done
8. I originally removed the imports because the line grew too big. I'll try to figure something out here.

In terms of capabilities and results/analysis -> all the tools are compatible with scikit-learn so people are free to measure what they'd like to use the metrics module that sklearn provides. I have a benchmark, one that is described in the course over [here](https://calmcode.io/model-mining/benchmark.html) and on our docs over [here](https://koaning.github.io/human-learn/examples/model-mining.html), but I don't consider the benchmark fair. I worry about listing benchmarks in the paper here mainly because the results may certainly vary depending on the dataset that one is working with.
