### 2021-08-01 #3448 @ahurriyetoglu
@koaning thanks!

I could not understand what do you mean by code formatting. I mean they should have a name and a mention in the text like the tables and figures.

A use case result could provide a sense of what to expect from your tool, although the exact results would change according to context. If you provide your context and your results, this will be enough.
