### 2021-08-01 #3448 @jbytecode
@ahurriyetoglu thank you for your comments. maybe you can give a code snippet here to clarify what you mean exactly, or, send a pull request to the main repo that implements the changes for paper.md.
