### 2021-08-01 #3448 @ahurriyetoglu
@jbytecode, @koaning I just say: using "from sklearn.metrics import *" is not a good practice. @koaning has already fixed it.
