### 2021-08-01 #3448 @koaning
@ahurriyetoglu I was referring to this bit:

> Code snippets require a numbering and in-text-mentioning as well.

I wouldn't mind adding line numbers to the code, same for in-text mentions, but I don't know if the markdown syntax allows for that. I couldn't find anything on [the docs](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography). Is there a guide on how to add that to JOSS papers?
