### 2021-08-02 #3448 @koaning
In terms of sharing some results, would [this result from the docs](https://koaning.github.io/human-learn/guide/function-classifier/function-classifier.html#comparison) be valid to share? It's on the titanic dataset, so it's not the best "benchmark dataset" out there, but it does demonstrate the validity of the approach.
