### 2021-08-06 #3448 @koaning
I just added the table with the comparison.
