### 2021-08-11 #3448 @jbytecode
Hi @koaning,

The reviewers think the software is valuable, but the article is a little bit far from being an academic article as in its current form. Please consider the suggestions below and do some more work on the article.

- In summary, Line 6, "This package contains..." maybe refactored as

"Human-Learn is a Python package that ..." so the first sentence describes what is the package.

- In summary, Line 7, "There are tools.." maybe refactored as

"The package contains some tools .... including?... for?...." so the second sentence describes what is the main functionality and what the package contains.

- In summary, jupyter must be Jupyter.

- In short, please refactor the summary, what the software is, what language is it written in, what is the functionality, similar to the current form of the summary, but in more descriptive way.

- In Statement of need, the sentence "There has been a transition from rule-based..." requires a citation.

- In line 16 and in remaining parts of the paper, sentences start with "we", but it is more convenient way is to transform these sentences to "The software (or paper) performs ..." so the paper is not dependent on the authors but it depends on the software.

- In Statement of Need, please be more descriptive, what is the previous work on this subject, what is the current status of the literature, and what is the gap that Human-Learn fills. Do not hesitate to add more citations just because you stand on the shoulders of giants.

- In Line 27, refactor the sentence something like "Human-Learn can be installed using
\# pip3 install human-learn
in the OS terminal.

- In line 39, there is a comparison table but it is not described well. Please describe and be clear on what are the columns, what are they showing, how to interpret them, etc.

- Use a different font for package, function, language names, for example, Python maybe `Python` using the `\texttt` LaTeX directive.

- Please re-check the reviewers (especially the ones @ahurriyetoglu did stressed) comments carefully.
