+++ .github/workflows/cron.yml
@@ -1,0 +1,27 @@
+name: Cron Test Dependencies
+
+on: [push]
+
+jobs:
+  cron:
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        python-version: [3.7]
+        os: [macos-10.15, ubuntu-latest, windows-latest]
+        pre-release-dependencies: ["--pre", ""]
+    steps:
+    - uses: actions/checkout@v2
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v1
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Install dependencies -
+      run: |
+        python -m pip install --upgrade pip
+        pip install ${{pre-release}} human-learn
+        pip freeze
+    - name: Test with pytest
+      run: |
+        pip install -e ".[test]"
+        pytest
