+++ .github/workflows/cron.yml
@@ -19,7 +19,7 @@ jobs:
     - name: Install dependencies -
       run: |
         python -m pip install --upgrade pip
-        pip install ${{pre-release}} human-learn
+        pip install ${{pre-release-dependencies}} human-learn
         pip freeze
     - name: Test with pytest
       run: |
