+++ tests/test_outliers/test_interactive_outlier.py
@@ -42,6 +42,8 @@ from tests.conftest import (
             "check_fit_score_takes_y",
             "check_dtype_object",
             "check_estimators_empty_data_messages",
+            "check_sample_weights_list",
+            "check_sample_weights_pandas_series",
         ],
     ),
 )
