+++ tests/test_classification/test_interactive_classifier.py
@@ -41,6 +41,8 @@ from tests.conftest import (
             "check_fit_score_takes_y",
             "check_dtype_object",
             "check_estimators_empty_data_messages",
+            "check_sample_weights_list",
+            "check_sample_weights_pandas_series",
         ],
     ),
 )
