+++ tests/test_transformers/test_interactive_preprocessor.py
@@ -28,6 +28,8 @@ from tests.conftest import (
             "check_fit1d",
             "check_methods_subset_invariance",
             "check_fit2d_predict1d",
+            "check_sample_weights_list",
+            "check_sample_weights_pandas_series",
         ],
     ),
 )
