### 2022-02-17 #3448 @jbytecode
@koaning, are you ready to pick up this submission again?
