### 2022-04-05 #3448 @jbytecode
Dear @koaning

On behalf of JOSS, I would like to express my deepest regrets. We withdraw the submission on your behalf in order to stop the review process for now. Whenever you decide to submit this paper again, please mention this thread so we can pick up where it was left as much as possible.

We wish everything will be better for you.

Best.
